truncation of wide columns
