+++ src/domain/settings/users/index.tsx
@@ -13,6 +13,7 @@ import PlusIcon from "../../../components/fundamentals/icons/plus-icon"
 import EditIcon from "../../../components/fundamentals/icons/edit-icon"
 import SidebarTeamMember from "../../../components/molecules/sidebar-team-member"
 import useMedusa from "../../../hooks/use-medusa"
+import UserTable from "../../../components/templates/user-table"
 
 type UserListElement = {
   entity: any
@@ -57,26 +58,9 @@ const Users: React.FC = () => {
           .then(inviteData => {
             setUsers(userData.users)
             setInvites(inviteData.invites)
-
-            setElements([
-              ...userData.users.map((user, i) => ({
-                entity: user,
-                entityType: "user",
-                tableElement: getUserTableRow(user, i),
-              })),
-              ...inviteData.invites.map((invite, i) => ({
-                entity: invite,
-                entityType: "invite",
-                tableElement: getInviteTableRow(invite, i),
-              })),
-            ])
           })
       })
   }, [shouldRefetch])
-
-  useEffect(() => {
-    setShownElements(elements)
-  }, [elements])
 
   const actionables = [
     {
@@ -89,191 +73,6 @@ const Users: React.FC = () => {
       ),
     },
   ]
-
-  const getUserTableRow = (user, index) => {
-    return (
-      <Table.Row
-        key={`user-${index}`}
-        color={"inherit"}
-        actions={[
-          {
-            label: "Edit User",
-            onClick: () => setSelectedUser(user),
-            icon: <EditIcon size={20} />,
-          },
-          {
-            label: "Remove User",
-            variant: "danger",
-            onClick: () => {
-              setDeleteUser(true)
-              setSelectedUser(user)
-            },
-            icon: <TrashIcon size={20} />,
-          },
-        ]}
-      >
-        <Table.Cell>
-          <SidebarTeamMember user={user} />
-        </Table.Cell>
-        <Table.Cell>{user.email}</Table.Cell>
-        <Table.Cell className="inter-small-semibold text-violet-60">
-          {user.role.charAt(0).toUpperCase()}
-          {user.role.slice(1)}
-        </Table.Cell>
-        <Table.Cell></Table.Cell>
-      </Table.Row>
-    )
-  }
-
-  const getInviteTableRow = (invite, index) => {
-    return (
-      <Table.Row
-        key={`invite-${index}`}
-        actions={[
-          {
-            label: "Resend Invitation",
-            onClick: () => {
-              Medusa.invites
-                .resend(invite.id)
-                .then(() => {
-                  toaster("Invitiation link has been resent", "success")
-                })
-                .then(() => triggerRefetch())
-            },
-            icon: <RefreshIcon size={20} />,
-          },
-          {
-            label: "Remove Invitation",
-            variant: "danger",
-            onClick: () => {
-              setSelectedInvite(invite)
-            },
-            icon: <TrashIcon size={20} />,
-          },
-        ]}
-      >
-        <Table.Cell className="text-grey-40">
-          <SidebarTeamMember user={{ email: invite.user_email }} />
-        </Table.Cell>
-        <Table.Cell className="text-grey-40">{invite.user_email}</Table.Cell>
-        <Table.Cell></Table.Cell>
-        <Table.Cell>
-          {new Date(invite?.expires_at) < new Date() ? (
-            <StatusDot title={"Expired"} variant={"danger"} />
-          ) : (
-            <StatusDot title={"Pending"} variant={"success"} />
-          )}
-        </Table.Cell>
-      </Table.Row>
-    )
-  }
-
-  const filteringOptions = [
-    {
-      title: "Team permissions",
-      options: [
-        {
-          title: "All",
-          count: elements.length,
-          onClick: () => setShownElements(elements),
-        },
-        {
-          title: "Member",
-          count: shownElements.filter(
-            e => e.entityType === "user" && e.entity.role === "member"
-          ).length,
-          onClick: () =>
-            setShownElements(
-              shownElements.filter(
-                e => e.entityType === "user" && e.entity.role === "member"
-              )
-            ),
-        },
-        {
-          title: "Admin",
-          count: shownElements.filter(
-            e => e.entityType === "user" && e.entity.role === "admin"
-          ).length,
-          onClick: () =>
-            setShownElements(
-              shownElements.filter(
-                e => e.entityType === "user" && e.entity.role === "admin"
-              )
-            ),
-        },
-        {
-          title: "No team permissions",
-          count: shownElements.filter(e => e.entityType === "invite").length,
-          onClick: () =>
-            setShownElements(
-              shownElements.filter(e => e.entityType === "invite")
-            ),
-        },
-      ],
-    },
-    {
-      title: "Status",
-      options: [
-        {
-          title: "All",
-          count: elements.length,
-          onClick: () => setShownElements(elements),
-        },
-        {
-          title: "Active",
-          count: shownElements.filter(e => e.entityType === "user").length,
-          onClick: () =>
-            setShownElements(
-              shownElements.filter(e => e.entityType === "user")
-            ),
-        },
-        {
-          title: "Pending",
-          count: shownElements.filter(
-            e =>
-              e.entityType === "invite" &&
-              getInviteStatus(e.entity) === "pending"
-          ).length,
-          onClick: () =>
-            setShownElements(
-              shownElements.filter(
-                e =>
-                  e.entityType === "invite" &&
-                  getInviteStatus(e.entity) === "pending"
-              )
-            ),
-        },
-        {
-          title: "Expired",
-          count: shownElements.filter(
-            e =>
-              e.entityType === "invite" &&
-              getInviteStatus(e.entity) === "expired"
-          ).length,
-          onClick: () =>
-            setShownElements(
-              shownElements.filter(
-                e =>
-                  e.entityType === "invite" &&
-                  getInviteStatus(e.entity) === "expired"
-              )
-            ),
-        },
-      ],
-    },
-  ]
-
-  const handleUserSearch = (term: string) => {
-    setShownElements(
-      elements.filter(
-        e =>
-          e.entity?.first_name?.includes(term) ||
-          e.entity?.last_name?.includes(term) ||
-          e.entity?.email?.includes(term) ||
-          e.entity?.user_email?.includes(term)
-      )
-    )
-  }
 
   return (
     <div className="w-full h-full pb-4">
@@ -289,59 +88,17 @@ const Users: React.FC = () => {
           actionables={actionables}
         >
           <div className="flex h-full justify-between flex-col pt-2">
-            <Table
-              filteringOptions={filteringOptions}
-              enableSearch
-              handleSearch={handleUserSearch}
-            >
-              <Table.Head>
-                <Table.HeadRow>
-                  <Table.HeadCell className="w-72">Name</Table.HeadCell>
-                  <Table.HeadCell className="w-80">Email</Table.HeadCell>
-                  <Table.HeadCell className="w-72">
-                    Team permissions
-                  </Table.HeadCell>
-                  <Table.HeadCell>Status</Table.HeadCell>
-                </Table.HeadRow>
-              </Table.Head>
-              <Table.Body>{shownElements.map(e => e.tableElement)}</Table.Body>
-            </Table>
+            <UserTable
+              users={users}
+              invites={invites}
+              triggerRefetch={triggerRefetch}
+            />
             <div className="inter-small-regular text-grey-50">
               {users.length} member
               {users.length === 1 ? "" : "s"}
             </div>
           </div>
-          {selectedUser &&
-            (deleteUser ? (
-              <DeletePrompt
-                text={"Are you sure you want to remove this user?"}
-                heading={"Remove user"}
-                onDelete={() =>
-                  Medusa.users
-                    .delete(selectedUser.id)
-                    .then(() => triggerRefetch())
-                }
-                handleClose={handleClose}
-              />
-            ) : (
-              <EditUser
-                handleClose={handleClose}
-                user={selectedUser}
-                onSubmit={triggerRefetch}
-              />
-            ))}
-          {selectedInvite && (
-            <DeletePrompt
-              text={"Are you sure you want to remove this invite?"}
-              heading={"Remove invite"}
-              onDelete={() =>
-                Medusa.invites
-                  .delete(selectedInvite.id)
-                  .then(() => triggerRefetch())
-              }
-              handleClose={handleClose}
-            />
-          )}
+
           {showInviteModal && (
             <InviteModal
               handleClose={() => {
